backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
numpy>=1.26.0
pillow>=10.0.0
python-dotenv==1.0.0
google-cloud-aiplatform>=1.38.0
psycopg2-binary==2.9.9
requests>=2.31.0
pgvector==0.2.4
geoalchemy2==0.14.3
pandas>=2.0.0
